+++ src/main/java/com/example/security/controller/dto/response/CategoryResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.security.dto.response;
+package com.example.security.controller.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
