+++ src/main/java/com/example/security/controller/dto/response/ProductResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.security.dto.request;
+package com.example.security.controller.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -7,10 +7,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
-public class ProductRequestDto {
+public class ProductResponseDto {
+
+    @Schema(description = "Id товара")
+    private int productId;
 
     @Schema(description = "Наименование товара")
     private String productTitle;
