+++ src/main/java/com/example/security/domain/entity/Image.java
@@ -1,4 +1,4 @@
-package com.example.security.entity;
+package com.example.security.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
