+++ src/main/java/com/example/security/domain/enums/Role.java
@@ -1,0 +1,5 @@
+package com.example.security.domain.enums;
+
+public enum Role {
+    ADMIN, USER
+}
