+++ src/main/java/com/example/security/domain/enums/Status.java
@@ -1,4 +1,4 @@
-package com.example.security.enums;
+package com.example.security.domain.enums;
 
 public enum Status {
     Принят, Оформлен, Ожидает, Получен
