+++ src/main/java/com/example/security/domain/repository/ImageRepository.java
@@ -1,0 +1,7 @@
+package com.example.security.domain.repository;
+
+import com.example.security.domain.entity.Image;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ImageRepository extends JpaRepository<Image, Integer> {
+}
