+++ src/main/java/com/example/security/domain/repository/OrderRepository.java
@@ -1,6 +1,6 @@
-package com.example.security.repository;
+package com.example.security.domain.repository;
 
-import com.example.security.entity.Order;
+import com.example.security.domain.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
