+++ src/main/java/com/example/security/domain/repository/PersonRepository.java
@@ -1,11 +1,8 @@
-package com.example.security.repository;
+package com.example.security.domain.repository;
 
-import com.example.security.entity.Person;
+import com.example.security.domain.entity.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface PersonRepository extends JpaRepository<Person, Long> {
