--- src/main/java/com/example/security/enums/Role.java
@@ -1,5 +1,0 @@
-package com.example.security.enums;
-
-public enum Role {
-    ADMIN, USER
-}
