+++ src/main/java/com/example/security/mapper/PersonMapper.java
@@ -1,7 +1,7 @@
 package com.example.security.mapper;
 
-import com.example.security.dto.response.PersonResponseDto;
-import com.example.security.entity.Person;
+import com.example.security.controller.dto.response.PersonResponseDto;
+import com.example.security.domain.entity.Person;
 import org.mapstruct.Mapper;
 
 import java.util.List;
